Fix wrong screen drawing by not counting ASCII color codes
The previous code assumed that it could count every byte in the current
line towards the length of the visible line.

With invisible characters, though, such as ASCII color codes, the count
was wrong and lead to the output to "climb up the screen" because the
overflow handling broke and too many newlines were emitted.

The fix here introduces a little "state switch" that flips on when we
come across an ESC character and turns to off when we come across 'm',
the character that terminates an ASCII color code.